The header and subheader are complete. Icons were added for social media. Main Navigation menu is complete. The page icon was added. Iconography was added.
The webpage icon was added with a 64x64 png file.

The header is complete with the social media bar and icons, the logo, the navigation menu and a space for a search box.

The subheader is complete with Title, Subtitle and highlights among those two.

The typography was added under typography.css.

All the changes were verified to work properly with @media queries.